refactor(submission): create SubmissionForm
- combination of both SubmissionEditForm and SubmissionEditStepForm
- remove orphaned PropTypes
